factorize Character
