+++ src/main/java/fr/imt/fa20/kangourou/character/Direction.java
@@ -1,0 +1,6 @@
+package fr.imt.fa20.kangourou.character;
+
+public enum Direction {
+	RIGHT,
+	LEFT
+}
