+++ src/main/java/fr/imt/fa20/kangourou/character/snake/SnakeState.java
@@ -1,0 +1,5 @@
+package fr.imt.fa20.kangourou.character.snake;
+
+public enum SnakeState {
+	WALKING, STANDING_BY, SUFFERING
+}
