--- src/main/java/fr/imt/fa20/kangourou/player/PlayerDirection.java
@@ -1,6 +1,0 @@
-package fr.imt.fa20.kangourou.player;
-
-public enum PlayerDirection {
-	RIGHT,
-	LEFT
-}
